+++ test/integration/embedders/test_mixedbread.py
@@ -2,8 +2,12 @@ import json
 import os
 from pathlib import Path
 
-from test.integration.embedders.utils import validate_embedding_output
+from test.integration.embedders.utils import validate_embedding_output, validate_raw_embedder
 from test.integration.utils import requires_env
+from unstructured_ingest.embed.mixedbreadai import (
+    MixedbreadAIEmbeddingConfig,
+    MixedbreadAIEmbeddingEncoder,
+)
 from unstructured_ingest.v2.processes.embedder import Embedder, EmbedderConfig
 
 API_KEY = "MXBAI_API_KEY"
@@ -25,3 +29,19 @@ def test_mixedbread_embedder(embedder_file: Path):
     with embedder_file.open("r") as f:
         original_elements = json.load(f)
     validate_embedding_output(original_elements=original_elements, output_elements=results)
+
+
+@requires_env(API_KEY)
+def test_raw_mixedbread_embedder(embedder_file: Path):
+    api_key = get_api_key()
+    embedder = MixedbreadAIEmbeddingEncoder(
+        config=MixedbreadAIEmbeddingConfig(
+            api_key=api_key,
+        )
+    )
+    validate_raw_embedder(
+        embedder=embedder,
+        embedder_file=embedder_file,
+        expected_dimensions=(1024,),
+        expected_is_unit_vector=False,
+    )