+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * **Move default API URL parameter value to serverless API**
 * **Databricks Delta Table destination to v2**
+* **Add check that access config always wrapped in Secret**
+* **Add togetherai embedder support**
 
 ### Fixes